+++ plugins/azure/azure.go
@@ -43,12 +43,12 @@ func Register(clusters *clusters.Clusters, plugins *plugin.Plugins, config Confi
 	var instances []*instance.Instance
 
 	for _, cfg := range config {
-		instance, err := instance.New(cfg)
+		inst, err := instance.New(cfg)
 		if err != nil {
 			log.WithError(err).WithFields(logrus.Fields{"name": cfg.Name}).Fatalf("Could not create Azure instance")
 		}
 
-		instances = append(instances, instance)
+		instances = append(instances, inst)
 
 		plugins.Append(plugin.Plugin{
 			Name:        cfg.Name,
@@ -72,6 +72,10 @@ func Register(clusters *clusters.Clusters, plugins *plugin.Plugins, config Confi
 			containerInstancesRouter.Get("/containergroup/details", router.getContainerGroup)
 			containerInstancesRouter.Get("/containergroup/logs", router.getContainerLogs)
 			containerInstancesRouter.Put("/containergroup/restart", router.restartContainerGroup)
+		})
+
+		r.Route("/costmanagement", func(costManagementRouter chi.Router) {
+			costManagementRouter.Get("/actualcost", router.getActualCost)
 		})
 
 		r.Route("/kubernetesservices", func(kubernetesServicesRouter chi.Router) {