+++ plugins/azure/pkg/instance/instance.go
@@ -2,11 +2,13 @@ package instance
 
 import (
 	"github.com/kobsio/kobs/plugins/azure/pkg/instance/containerinstances"
+	"github.com/kobsio/kobs/plugins/azure/pkg/instance/costmanagement"
 	"github.com/kobsio/kobs/plugins/azure/pkg/instance/kubernetesservices"
 	"github.com/kobsio/kobs/plugins/azure/pkg/instance/monitor"
 	"github.com/kobsio/kobs/plugins/azure/pkg/instance/resourcegroups"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
+	"github.com/Azure/go-autorest/autorest/azure/auth"
 	"github.com/sirupsen/logrus"
 )
 
@@ -38,6 +40,7 @@ type Instance struct {
 	ResourceGroups     *resourcegroups.Client
 	KubernetesServices *kubernetesservices.Client
 	ContainerInstances *containerinstances.Client
+	CostManagement     *costmanagement.Client
 	Monitor            *monitor.Client
 }
 
@@ -48,10 +51,17 @@ func New(config Config) (*Instance, error) {
 		return nil, err
 	}
 
+	authorizerClientCredentialsConfig := auth.NewClientCredentialsConfig(config.Credentials.ClientID, config.Credentials.ClientSecret, config.Credentials.TenantID)
+	authorizer, err := authorizerClientCredentialsConfig.Authorizer()
+	if err != nil {
+		return nil, err
+	}
+
 	resourceGroups := resourcegroups.New(config.Credentials.SubscriptionID, credentials)
 	kubernetesServices := kubernetesservices.New(config.Credentials.SubscriptionID, credentials)
 	containerInstances := containerinstances.New(config.Credentials.SubscriptionID, credentials)
 	monitor := monitor.New(config.Credentials.SubscriptionID, credentials)
+	costManagement := costmanagement.New(config.Credentials.SubscriptionID, authorizer)
 
 	return &Instance{
 		Name:               config.Name,
@@ -59,6 +69,7 @@ func New(config Config) (*Instance, error) {
 		ResourceGroups:     resourceGroups,
 		KubernetesServices: kubernetesServices,
 		ContainerInstances: containerInstances,
+		CostManagement:     costManagement,
 		Monitor:            monitor,
 	}, nil
 }