+++ plugins/azure/src/utils/helpers.ts
@@ -1,5 +1,6 @@
 import { Serie } from '@nivo/line';
 
+import { IPieDatum, IQueryResult } from '../components/costmanagement/interfaces';
 import { IMetric } from './interfaces';
 import { formatTime as formatTimeCore } from '@kobsio/plugin-core';
 
@@ -57,6 +58,21 @@ export const convertMetric = (metric: IMetric): Serie[] => {
   }
 
   return series;
+};
+
+// convertQueryResult returns the cost management query result in the format required for nivo pie canvas.
+export const convertQueryResult = (data: IQueryResult): IPieDatum[] => {
+  const pieData: IPieDatum[] = [];
+
+  for (let i = 0; i < data.properties.rows.length; i++) {
+    pieData.push({
+      id: data.properties.rows[i][1],
+      label: data.properties.rows[i][1],
+      value: data.properties.rows[i][0],
+    });
+  }
+
+  return pieData;
 };
 
 // formatMetric is used to auto format the values of a metric. When the unit of a metric is "Bytes" we auto format the